% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geoms.R
\name{extract_geoms}
\alias{extract_geoms}
\title{Given a Letter Find Geoms Inside a Package}
\usage{
extract_geoms(pkg, prefix = "^geom_")
}
\arguments{
\item{pkg}{A character string or unquoted package name. The package from which
to extract functions.}

\item{prefix}{A character string regex pattern used to filter function names.
Defaults to \code{"^geom_"}.}
}
\value{
A character vector containing the names of the functions that match
the prefix inside the given package.
}
\description{
This function extracts the names of all functions inside a specified package
that begin with a given prefix (e.g., \code{"geom_"}). It is useful for quickly
identifying available \code{geom_} functions in packages like \strong{ggplot2} or \strong{ggalt}.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  extract_geoms(ggplot2, "^geom_s")
}
}
