% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_encyclopedia.R
\name{load_encyclopedia}
\alias{load_encyclopedia}
\title{Load Encyclopedia}
\usage{
load_encyclopedia()
}
\value{
Opens Data Visualisation Geometries Encyclopedia Book on your browser
}
\description{
This function loads Data Visualisation Geometries Encyclopedia Book
}
