% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldbankdata.R
\docType{data}
\name{worldbankdata}
\alias{worldbankdata}
\title{worldbankdata}
\format{
A data frame with 7937 rows and 7 variables.
}
\source{
The World Bank Data Catalogue (accessed on 1 Dec 2023)
}
\usage{
data(worldbankdata)
}
\description{
This dataset provides details about countries based on information from the World Bank.
}
\examples{
data(worldbankdata)
head(worldbankdata)
summary(worldbankdata)
}
\keyword{datasets}
