% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldHappinessScore.R
\docType{data}
\name{WorldHappinessScore}
\alias{WorldHappinessScore}
\title{World Happiness Scores (2022–2024)}
\format{
A data frame with 145 rows and 6 variables:
\describe{
\item{flagCode}{Country flag code.}
\item{country}{Country name.}
\item{WorldHappinessScore_2024}{World happiness score for the year 2024.}
\item{WorldHappinessScore_2023}{World happiness score for the year 2023.}
\item{WorldHappinessScore_2022}{World happiness score for the year 2022.}
}
}
\source{
Data accessed from World Population Review:
\url{https://worldpopulationreview.com/country-rankings/happiest-countries-in-the-world}
(accessed on 3 January 2026).
}
\usage{
data(WorldHappinessScore)
}
\description{
This dataset provides World Happiness Scores for the years 2022, 2023,
and 2024. The World Happiness Score is measured on a 0–10 scale,
where 0 represents the worst possible life and 10 represents the
best possible life.
}
\keyword{datasets}
